Correct computation of progress (integer round off bug).

+++ src/net/sf/ij_plugins/im3d/grow/SRG.java
@@ -230,7 +230,7 @@ public final class SRG extends DefaultProgressReporter {
 
         // Calculate increment, make sure that different/larger than 0 otherwise '%' operation will fail.
         final long progressIncrement = Math.max(pixelsToProcess / 25, 1);
-        this.notifyProgressListeners(processedPixelCount / pixelsToProcess);
+        this.notifyProgressListeners(processedPixelCount / (double) pixelsToProcess);
 
 
         // Process candidates
